+++ Assets/Mochineko/FacialExpressions.Samples/SampleForVoiceVoxAndVRM.cs
@@ -1,5 +1,6 @@
 #nullable enable
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using Cysharp.Threading.Tasks;
@@ -25,17 +26,32 @@ namespace Mochineko.FacialExpressions.Samples
     // ReSharper disable once InconsistentNaming
     internal sealed class SampleForVoiceVoxAndVRM : MonoBehaviour
     {
-        [SerializeField] private string path = string.Empty;
-        [SerializeField] private string text = string.Empty;
-        [SerializeField] private int speakerID;
-        [SerializeField] private AudioSource? audioSource = null;
-        [SerializeField] private bool skipSpeechSynthesis = false;
-        [SerializeField] private BasicEmotion basicEmotion = BasicEmotion.Neutral;
-        [SerializeField] private float emotionWeight = 1f;
-        [SerializeField] private float emotionFollowingTime = 1f;
-        
-        private ILipAnimator? lipAnimator;
-        private IEyelidAnimator? eyelidAnimator;
+        [SerializeField]
+        private string path = string.Empty;
+
+        [SerializeField]
+        private string text = string.Empty;
+
+        [SerializeField]
+        private int speakerID;
+
+        [SerializeField]
+        private AudioSource? audioSource = null;
+
+        [SerializeField]
+        private bool skipSpeechSynthesis = false;
+
+        [SerializeField]
+        private BasicEmotion basicEmotion = BasicEmotion.Neutral;
+
+        [SerializeField]
+        private float emotionWeight = 1f;
+
+        [SerializeField]
+        private float emotionFollowingTime = 1f;
+
+        private FollowingLipAnimator? lipAnimator;
+        private IDisposable? eyelidAnimationLoop;
         private ExclusiveFollowingEmotionAnimator<BasicEmotion>? emotionAnimator;
         private AudioClip? audioClip;
 
@@ -53,7 +69,7 @@ namespace Mochineko.FacialExpressions.Samples
         private async void Start()
         {
             VoiceVoxBaseURL.BaseURL = "http://127.0.0.1:50021";
-            
+
             var binary = await File.ReadAllBytesAsync(
                 path,
                 this.GetCancellationTokenOnDestroy());
@@ -65,20 +81,28 @@ namespace Mochineko.FacialExpressions.Samples
             var lipMorpher = new VRMLipMorpher(instance.Runtime.Expression);
             lipAnimator = new FollowingLipAnimator(lipMorpher);
 
+            var eyelidFrames = ProbabilisticEyelidAnimationGenerator
+                .Generate(
+                    Eyelid.Both,
+                    blinkCount: 20
+                );
+
             var eyelidMorpher = new VRMEyelidMorpher(instance.Runtime.Expression);
-            eyelidAnimator = new SequentialEyelidAnimator(eyelidMorpher);
+            var eyelidAnimator = new SequentialEyelidAnimator(eyelidMorpher);
+            eyelidAnimationLoop = new LoopEyelidAnimator(eyelidAnimator, eyelidFrames);
 
-            var eyelidFrames = ProbabilisticEyelidAnimationGenerator.Generate(
-                Eyelid.Both,
-                blinkCount: 20);
-
-            eyelidAnimator.AnimateAsync(
-                    eyelidFrames,
-                    loop: true,
-                    this.GetCancellationTokenOnDestroy())
-                .Forget();
-            
-            var emotionMorpher = new VRMEmotionMorpher(instance.Runtime.Expression);
+            var emotionMorpher = new VRMEmotionMorpher<BasicEmotion>(
+                instance.Runtime.Expression,
+                keyMap: new Dictionary<BasicEmotion, ExpressionKey>
+                {
+                    [BasicEmotion.Neutral] = ExpressionKey.Neutral,
+                    [BasicEmotion.Happy] = ExpressionKey.Happy,
+                    [BasicEmotion.Sad] = ExpressionKey.Sad,
+                    [BasicEmotion.Angry] = ExpressionKey.Angry,
+                    [BasicEmotion.Fearful] = ExpressionKey.Neutral,
+                    [BasicEmotion.Surprised] = ExpressionKey.Surprised,
+                    [BasicEmotion.Disgusted] = ExpressionKey.Neutral,
+                });
             emotionAnimator = new ExclusiveFollowingEmotionAnimator<BasicEmotion>(
                 emotionMorpher,
                 followingTime: emotionFollowingTime);
@@ -86,7 +110,6 @@ namespace Mochineko.FacialExpressions.Samples
 
         private void Update()
         {
-            eyelidAnimator?.Update();
             lipAnimator?.Update();
             emotionAnimator?.Update();
         }
@@ -98,6 +121,7 @@ namespace Mochineko.FacialExpressions.Samples
                 Destroy(audioClip);
                 audioClip = null;
             }
+            eyelidAnimationLoop?.Dispose();
         }
 
         private static async UniTask<Vrm10Instance> LoadVRMAsync(
@@ -139,6 +163,11 @@ namespace Mochineko.FacialExpressions.Samples
                 return;
             }
 
+            if (lipAnimator == null)
+            {
+                return;
+            }
+
             if (audioClip != null)
             {
                 Destroy(audioClip);
@@ -149,70 +178,27 @@ namespace Mochineko.FacialExpressions.Samples
 
             await UniTask.SwitchToThreadPool();
 
-            AudioQuery audioQuery;
-            var createQueryResult = await queryCreationPolicy.ExecuteAsync(
-                async innerCancellationToken => await QueryCreationAPI.CreateQueryAsync(
-                    HttpClientPool.PooledClient,
-                    text: text,
-                    speaker: speakerID,
-                    coreVersion: null,
-                    cancellationToken: innerCancellationToken),
-                cancellationToken);
-            if (createQueryResult is IUncertainSuccessResult<AudioQuery> createQuerySuccess)
-            {
-                audioQuery = createQuerySuccess.Result;
-                Debug.Log(
-                    $"[VOICEVOX_API.Samples] Succeeded to create query from text:{text} -> {JsonConvert.SerializeObject(audioQuery)}.");
-            }
-            else if (createQueryResult is IUncertainRetryableResult<AudioQuery> createQueryRetryable)
-            {
-                Debug.LogError(
-                    $"[VOICEVOX_API.Samples] Failed to create query because -> {createQueryRetryable.Message}.");
-                return;
-            }
-            else if (createQueryResult is IUncertainFailureResult<AudioQuery> createQueryFailure)
-            {
-                Debug.LogError(
-                    $"[VOICEVOX_API.Samples] Failed to create query because -> {createQueryFailure.Message}.");
-                return;
-            }
-            else
-            {
-                throw new UncertainResultPatternMatchException(nameof(createQueryResult));
-            }
+            var audioQuery = (await queryCreationPolicy.ExecuteAsync(
+                    async innerCancellationToken => await QueryCreationAPI.CreateQueryAsync(
+                        HttpClientPool.PooledClient,
+                        text: text,
+                        speaker: speakerID,
+                        coreVersion: null,
+                        cancellationToken: innerCancellationToken),
+                    cancellationToken)
+                ).Unwrap();
 
             // Synthesize speech from AudioQuery by VOICEVOX synthesis API.
-            Stream stream;
-            var synthesisResult = await synthesisPolicy.ExecuteAsync(
-                async innerCancellationToken => await SynthesisAPI.SynthesizeAsync(
-                    HttpClientPool.PooledClient,
-                    audioQuery: audioQuery,
-                    speaker: speakerID,
-                    enableInterrogativeUpspeak: null,
-                    coreVersion: null,
-                    cancellationToken: innerCancellationToken),
-                cancellationToken);
-            if (synthesisResult is IUncertainSuccessResult<Stream> synthesisSuccess)
-            {
-                stream = synthesisSuccess.Result;
-                Debug.Log($"[VOICEVOX_API.Samples] Succeeded to synthesis speech from text:{text}.");
-            }
-            else if (synthesisResult is IUncertainRetryableResult<Stream> synthesisRetryable)
-            {
-                Debug.LogError(
-                    $"[VOICEVOX_API.Samples] Failed to synthesis speech because -> {synthesisRetryable.Message}.");
-                return;
-            }
-            else if (synthesisResult is IUncertainFailureResult<Stream> synthesisFailure)
-            {
-                Debug.LogError(
-                    $"[VOICEVOX_API.Samples] Failed to synthesis speech because -> {synthesisFailure.Message}.");
-                return;
-            }
-            else
-            {
-                throw new UncertainResultPatternMatchException(nameof(synthesisResult));
-            }
+            var stream = (await synthesisPolicy.ExecuteAsync(
+                    async innerCancellationToken => await SynthesisAPI.SynthesizeAsync(
+                        HttpClientPool.PooledClient,
+                        audioQuery: audioQuery,
+                        speaker: speakerID,
+                        enableInterrogativeUpspeak: null,
+                        coreVersion: null,
+                        cancellationToken: innerCancellationToken),
+                    cancellationToken)
+                ).Unwrap();
 
             if (!skipSpeechSynthesis)
             {
@@ -239,25 +225,26 @@ namespace Mochineko.FacialExpressions.Samples
                 {
                     await stream.DisposeAsync();
                 }
-                
+
                 await UniTask.SwitchToMainThread(cancellationToken);
 
                 // Play AudioClip.
                 audioSource.clip = audioClip;
                 audioSource.PlayDelayed(0.1f);
             }
-            
-            var lipFrames = AudioQueryConverter.ConvertToSequentialAnimationFrames(audioQuery);
+
+            var lipFrames = AudioQueryConverter
+                .ConvertToSequentialAnimationFrames(audioQuery);
 
             await UniTask.Delay(
                 TimeSpan.FromSeconds(0.1f),
-                cancellationToken:cancellationToken);
-            
+                cancellationToken: cancellationToken);
+
             lipAnimator
-                ?.AnimateAsync(lipFrames, cancellationToken)
+                .SetTargetAsync(lipFrames, cancellationToken)
                 .Forget();
         }
-        
+
         [ContextMenu(nameof(Emote))]
         public void Emote()
         {
@@ -267,4 +254,6 @@ namespace Mochineko.FacialExpressions.Samples
                     weight: emotionWeight));
         }
     }
+
+    // ReSharper disable once InconsistentNaming
 }