+++ Assets/Mochineko/FacialExpressions.Samples/VolumeBasedLipSyncSample.cs
@@ -1,5 +1,6 @@
 #nullable enable
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using Cysharp.Threading.Tasks;
@@ -47,15 +48,15 @@ namespace Mochineko.FacialExpressions.Samples
 
         [SerializeField]
         private float emotionFollowingTime = 1f;
-        
+
         [SerializeField]
         private float volumeSmoothTime = 0.1f;
-        
+
         [SerializeField]
         private float volumeMultiplier = 1f;
 
         private VolumeBasedLipAnimator? lipAnimator;
-        private IEyelidAnimator? eyelidAnimator;
+        private IDisposable? eyelidAnimationLoop;
         private ExclusiveFollowingEmotionAnimator<BasicEmotion>? emotionAnimator;
         private AudioClip? audioClip;
 
@@ -96,20 +97,26 @@ namespace Mochineko.FacialExpressions.Samples
                 volumeMultiplier: volumeMultiplier,
                 samplesCount: 1024);
 
-            var eyelidMorpher = new VRMEyelidMorpher(instance.Runtime.Expression);
-            eyelidAnimator = new SequentialEyelidAnimator(eyelidMorpher);
-
             var eyelidFrames = ProbabilisticEyelidAnimationGenerator.Generate(
                 Eyelid.Both,
                 blinkCount: 20);
 
-            eyelidAnimator.AnimateAsync(
-                    eyelidFrames,
-                    loop: true,
-                    this.GetCancellationTokenOnDestroy())
-                .Forget();
+            var eyelidMorpher = new VRMEyelidMorpher(instance.Runtime.Expression);
+            var eyelidAnimator = new SequentialEyelidAnimator(eyelidMorpher);
+            eyelidAnimationLoop = new LoopEyelidAnimator(eyelidAnimator, eyelidFrames);
 
-            var emotionMorpher = new VRMEmotionMorpher(instance.Runtime.Expression);
+            var emotionMorpher = new VRMEmotionMorpher<BasicEmotion>(
+                instance.Runtime.Expression,
+                keyMap: new Dictionary<BasicEmotion, ExpressionKey>
+                {
+                    [BasicEmotion.Neutral] = ExpressionKey.Neutral,
+                    [BasicEmotion.Happy] = ExpressionKey.Happy,
+                    [BasicEmotion.Sad] = ExpressionKey.Sad,
+                    [BasicEmotion.Angry] = ExpressionKey.Angry,
+                    [BasicEmotion.Fearful] = ExpressionKey.Neutral,
+                    [BasicEmotion.Surprised] = ExpressionKey.Surprised,
+                    [BasicEmotion.Disgusted] = ExpressionKey.Neutral,
+                });
             emotionAnimator = new ExclusiveFollowingEmotionAnimator<BasicEmotion>(
                 emotionMorpher,
                 followingTime: emotionFollowingTime);
@@ -117,7 +124,6 @@ namespace Mochineko.FacialExpressions.Samples
 
         private void Update()
         {
-            eyelidAnimator?.Update();
             lipAnimator?.Update();
             emotionAnimator?.Update();
         }
@@ -129,6 +135,7 @@ namespace Mochineko.FacialExpressions.Samples
                 Destroy(audioClip);
                 audioClip = null;
             }
+            eyelidAnimationLoop?.Dispose();
         }
 
         private static async UniTask<Vrm10Instance> LoadVRMAsync(