+++ Assets/Mochineko/FacialExpressions/Blink/EyelidAnimationFrame.cs
@@ -5,7 +5,7 @@ using System.Runtime.InteropServices;
 namespace Mochineko.FacialExpressions.Blink
 {
     /// <summary>
-    /// Frame of eyelid animation.
+    /// A frame of eyelid animation.
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
     public readonly struct EyelidAnimationFrame : IEquatable<EyelidAnimationFrame>
@@ -19,13 +19,21 @@ namespace Mochineko.FacialExpressions.Blink
         /// </summary>
         public readonly float durationSeconds;
 
+        /// <summary>
+        /// Creates a new instance of <see cref="EyelidAnimationFrame"/>.
+        /// </summary>
+        /// <param name="sample">Sample of eyelid morphing at frame.</param>
+        /// <param name="durationSeconds">Duration time of frame in seconds.</param>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         public EyelidAnimationFrame(EyelidSample sample, float durationSeconds)
         {
             if (durationSeconds < 0f)
             {
-                throw new ArgumentOutOfRangeException(nameof(durationSeconds));
+                throw new ArgumentOutOfRangeException(
+                    nameof(durationSeconds), durationSeconds,
+                    "Duration time must be greater than or equal to 0.");
             }
-            
+
             this.sample = sample;
             this.durationSeconds = durationSeconds;
         }