+++ Assets/Mochineko/FacialExpressions/Blink/SkinnedMeshEyelidMorpher.cs
@@ -11,20 +11,40 @@ namespace Mochineko.FacialExpressions.Blink
     {
         private readonly SkinnedMeshRenderer skinnedMeshRenderer;
         private readonly IReadOnlyDictionary<Eyelid, int> indexMap;
+        private readonly bool separateBoth;
 
+        /// <summary>
+        /// Creates a new instance of <see cref="SkinnedMeshEyelidMorpher"/>.
+        /// </summary>
+        /// <param name="skinnedMeshRenderer">Target renderer.</param>
+        /// <param name="indexMap">Map of eyelid and blend shape index.</param>
+        /// <param name="separateBoth">Whether separate both eyelids blend shape.</param>
         public SkinnedMeshEyelidMorpher(
             SkinnedMeshRenderer skinnedMeshRenderer,
-            IReadOnlyDictionary<Eyelid, int> indexMap)
+            IReadOnlyDictionary<Eyelid, int> indexMap,
+            bool separateBoth = false)
         {
             this.skinnedMeshRenderer = skinnedMeshRenderer;
             this.indexMap = indexMap;
+            this.separateBoth = separateBoth;
         }
 
         public void MorphInto(EyelidSample sample)
         {
-            if (indexMap.TryGetValue(sample.eyelid, out var index))
+            if (separateBoth && sample.eyelid == Eyelid.Both)
             {
-                skinnedMeshRenderer.SetBlendShapeWeight(index, sample.weight);
+                if (indexMap.TryGetValue(Eyelid.Left, out var rightIndex))
+                {
+                    skinnedMeshRenderer.SetBlendShapeWeight(rightIndex, sample.weight * 100f);
+                }
+                if (indexMap.TryGetValue(Eyelid.Right, out var leftIndex))
+                {
+                    skinnedMeshRenderer.SetBlendShapeWeight(leftIndex, sample.weight * 100f);
+                }
+            }
+            else if (indexMap.TryGetValue(sample.eyelid, out var index))
+            {
+                skinnedMeshRenderer.SetBlendShapeWeight(index, sample.weight * 100f);
             }
         }
 
@@ -32,7 +52,7 @@ namespace Mochineko.FacialExpressions.Blink
         {
             if (indexMap.TryGetValue(eyelid, out var index))
             {
-                return skinnedMeshRenderer.GetBlendShapeWeight(index);
+                return skinnedMeshRenderer.GetBlendShapeWeight(index) / 100f;
             }
             else
             {
@@ -42,9 +62,10 @@ namespace Mochineko.FacialExpressions.Blink
 
         public void Reset()
         {
-            MorphInto(new EyelidSample(Eyelid.Both, 0f));
-            MorphInto(new EyelidSample(Eyelid.Left, 0f));
-            MorphInto(new EyelidSample(Eyelid.Right, 0f));
+            foreach (var pair in indexMap)
+            {
+                skinnedMeshRenderer.SetBlendShapeWeight(pair.Value, 0f);
+            }
         }
     }
 }