+++ Assets/Mochineko/FacialExpressions/Emotion/ExclusiveFollowingEmotionAnimator.cs
@@ -6,10 +6,10 @@ using UnityEngine;
 namespace Mochineko.FacialExpressions.Emotion
 {
     /// <summary>
-    /// An emotion animator that animates emotion by following target weights.
+    /// An emotion animator that animates emotion by following target weights exclusively.
     /// </summary>
     public sealed class ExclusiveFollowingEmotionAnimator<TEmotion>
-        : IEmotionAnimator<TEmotion>
+        : IFramewiseEmotionAnimator<TEmotion>
         where TEmotion : Enum
     {
         private readonly IEmotionMorpher<TEmotion> morpher;
@@ -17,15 +17,32 @@ namespace Mochineko.FacialExpressions.Emotion
         private readonly Dictionary<TEmotion, EmotionSample<TEmotion>> targets = new();
         private readonly Dictionary<TEmotion, float> velocities = new();
 
+        /// <summary>
+        /// Creates a new instance of <see cref="ExclusiveFollowingEmotionAnimator{TEmotion}"/>.
+        /// </summary>
+        /// <param name="morpher">Target morpher.</param>
+        /// <param name="followingTime">Following time to smooth dump.</param>
+        /// <exception cref="ArgumentOutOfRangeException"></exception>
         public ExclusiveFollowingEmotionAnimator(
             IEmotionMorpher<TEmotion> morpher,
             float followingTime)
         {
+            if (followingTime <= 0f)
+            {
+                throw new ArgumentOutOfRangeException(
+                    nameof(followingTime), followingTime,
+                    "Following time must be greater than 0.");
+            }
+
             this.morpher = morpher;
             this.followingTime = followingTime;
             this.morpher.Reset();
         }
 
+        /// <summary>
+        /// Emotes.
+        /// </summary>
+        /// <param name="sample"></param>
         public void Emote(EmotionSample<TEmotion> sample)
         {
             if (targets.ContainsKey(sample.emotion))