+++ Assets/Mochineko/FacialExpressions/LipSync/FollowingLipAnimator.cs
@@ -9,9 +9,9 @@ using UnityEngine;
 namespace Mochineko.FacialExpressions.LipSync
 {
     /// <summary>
-    /// A lip animator that animates lip by following target weights.
+    /// A framewise lip animator that animates lip by following target weights.
     /// </summary>
-    public sealed class FollowingLipAnimator : ILipAnimator
+    public sealed class FollowingLipAnimator : IFramewiseLipAnimator
     {
         private readonly ILipMorpher morpher;
         private readonly float dt;
@@ -22,17 +22,39 @@ namespace Mochineko.FacialExpressions.LipSync
 
         private CancellationTokenSource? animationCanceller;
 
+        /// <summary>
+        /// Creates a new instance of <see cref="FollowingLipAnimator"/>.
+        /// </summary>
+        /// <param name="morpher">Target morpher.</param>
+        /// <param name="initialFollowingVelocity">Initial velocity of smooth damp.</param>
+        /// <param name="followingTime">Time of smooth damp.</param>
         public FollowingLipAnimator(
             ILipMorpher morpher,
             float initialFollowingVelocity = 0.1f,
             float followingTime = 0.005f)
         {
+            if (initialFollowingVelocity < 0f)
+            {
+                throw new ArgumentOutOfRangeException(
+                    nameof(initialFollowingVelocity),
+                    initialFollowingVelocity,
+                    "Initial velocity must be greater than or equal to 0.");
+            }
+
+            if (followingTime < 0f)
+            {
+                throw new ArgumentOutOfRangeException(
+                    nameof(followingTime),
+                    followingTime,
+                    "Following time must be greater than or equal to 0.");
+            }
+
             this.morpher = morpher;
             this.initialFollowingVelocity = initialFollowingVelocity;
             this.followingTime = followingTime;
         }
 
-        public async UniTask AnimateAsync(
+        public async UniTask SetTargetAsync(
             IEnumerable<LipAnimationFrame> frames,
             CancellationToken cancellationToken)
         {
@@ -70,6 +92,19 @@ namespace Mochineko.FacialExpressions.LipSync
 
             animationCanceller?.Cancel();
             animationCanceller = null;
+        }
+
+        public void SetTarget(LipSample sample)
+        {
+            if (targetWeights.ContainsKey(sample.viseme))
+            {
+                targetWeights[sample.viseme] = sample.weight;
+            }
+            else
+            {
+                targetWeights.Add(sample.viseme, sample.weight);
+                followingVelocities.Add(sample.viseme, initialFollowingVelocity);
+            }
         }
 
         public void Update()