+++ Assets/Mochineko/FacialExpressions/LipSync/SequentialLipAnimator.cs
@@ -8,15 +8,18 @@ using Mochineko.Relent.Extensions.UniTask;
 namespace Mochineko.FacialExpressions.LipSync
 {
     /// <summary>
-    /// A lip animator that animates lip sequentially.
+    /// A sequential lip animator.
     /// </summary>
-    public sealed class SequentialLipAnimator : ILipAnimator
+    public sealed class SequentialLipAnimator : ISequentialLipAnimator
     {
         private readonly ILipMorpher morpher;
 
-        private LipSample currentSample = new(Viseme.sil, 0f);
         private CancellationTokenSource? animationCanceller;
 
+        /// <summary>
+        /// Creates a new instance of <see cref="SequentialLipAnimator"/>.
+        /// </summary>
+        /// <param name="morpher">Target morpher.</param>
         public SequentialLipAnimator(ILipMorpher morpher)
         {
             this.morpher = morpher;
@@ -29,7 +32,7 @@ namespace Mochineko.FacialExpressions.LipSync
             animationCanceller?.Cancel();
             animationCanceller = CancellationTokenSource
                 .CreateLinkedTokenSource(cancellationToken);
-            
+
             morpher.Reset();
 
             foreach (var frame in frames)
@@ -39,33 +42,22 @@ namespace Mochineko.FacialExpressions.LipSync
                     break;
                 }
 
-                currentSample = frame.sample;
+                morpher.MorphInto(frame.sample);
 
                 var result = await RelentUniTask.Delay(
                     TimeSpan.FromSeconds(frame.durationSeconds),
                     cancellationToken: animationCanceller.Token);
-                
+
                 // Cancelled
                 if (result.Failure)
                 {
                     break;
                 }
             }
-            
+
             morpher.Reset();
-            currentSample = new LipSample(Viseme.sil, 0f);
 
             animationCanceller = null;
-        }
-
-        public void Update()
-        {
-            morpher.MorphInto(currentSample);
-        }
-
-        public void Reset()
-        {
-            morpher.Reset();
         }
     }
 }